Rename middleware restricted
